+++ ui/src/services/segmenter/SegmenterScope.js
@@ -1,0 +1,13 @@
+export const getSegmenterScope = (scope) => {
+  const status = {
+    project: {
+      label: "Project",
+      color: "primary",
+    },
+    global: {
+      label: "Global",
+      color: "secondary",
+    },
+  };
+  return status[scope];
+};
